+++ server/src/covet/api/__init__.py
@@ -6,6 +6,7 @@ from covet.api import auth as auth_router
 from covet.api import collections as collections_router
 from covet.api import contacts as contacts_router
 from covet.api import imports as imports_router
+from covet.api import invitations as invitations_router
 from covet.api import items as items_router
 from covet.api import loans as loans_router
 from covet.api import meta as meta_router
@@ -24,5 +25,6 @@ api_router.include_router(loans_router.router)
 api_router.include_router(sync_router.router)
 api_router.include_router(imports_router.router)
 api_router.include_router(share_router.router)
+api_router.include_router(invitations_router.router)
 
 __all__ = ["api_router"]